Settled down all db related packages into db package
