+++ src/main/java/kg/peaksoft/ebookb4/api/AuthController.java
@@ -6,8 +6,8 @@ import kg.peaksoft.ebookb4.dto.request.SignupRequestVendor;
 import kg.peaksoft.ebookb4.dto.response.JwtResponse;
 import kg.peaksoft.ebookb4.security.jwt.JwtUtils;
 import kg.peaksoft.ebookb4.security.services.UserDetailsImpl;
-import kg.peaksoft.ebookb4.service.ClientService;
-import kg.peaksoft.ebookb4.service.VendorService;
+import kg.peaksoft.ebookb4.db.service.ClientService;
+import kg.peaksoft.ebookb4.db.service.VendorService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
