+++ src/main/java/kg/peaksoft/ebookb4/db/models/bookClasses/ElectronicBook.java
@@ -1,7 +1,5 @@
-package kg.peaksoft.ebookb4.models.others;
+package kg.peaksoft.ebookb4.db.models.bookClasses;
 
-import kg.peaksoft.ebookb4.models.bookClasses.Book;
-import kg.peaksoft.ebookb4.models.userClasses.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,24 +7,24 @@ import lombok.Setter;
 
 import javax.persistence.*;
 
+/**
+ * Author: Zhanarbek Abdurasulov
+ * Date: 12/4/22
+ */
 @Entity
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Favorites {
+public class ElectronicBook {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "generator")
     @SequenceGenerator(name = "generator", sequenceName = "generator", allocationSize = 1)
-    @Column(name = "favorite_id")
-    private Long favoriteId;
+    private Long ebookId;
+    private String publishingHouse;
+    private String fragmentOfBook;
+    private Integer numberOfPages;
+    private String urlOfBookFromCloud;
 
-    @ManyToOne
-    @JoinColumn(name = "book_id")
-    private Book book;
-
-    @OneToOne
-    @JoinColumn(name = "user_id")
-    private User User;
 
 }
