+++ src/main/java/kg/peaksoft/ebookb4/db/models/bookClasses/PaperBook.java
@@ -1,32 +1,26 @@
-package kg.peaksoft.ebookb4.models.bookClasses;
+package kg.peaksoft.ebookb4.db.models.bookClasses;
 
-import kg.peaksoft.ebookb4.models.others.FileSources;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.io.File;
 
-/**
- * Author: Zhanarbek Abdurasulov
- * Date: 12/4/22
- */
 @Entity
-@Getter
-@Setter
+@Getter @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ElectronicBook {
+public class PaperBook {
+
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "generator")
     @SequenceGenerator(name = "generator", sequenceName = "generator", allocationSize = 1)
-    private Long ebookId;
+    private Long paperBookId;
+
     private String publishingHouse;
     private String fragmentOfBook;
     private Integer numberOfPages;
-    private String urlOfBookFromCloud;
-
+    private Integer numberOfSelected;
 
 }
