+++ src/main/java/kg/peaksoft/ebookb4/db/models/enums/Language.java
@@ -1,4 +1,4 @@
-package kg.peaksoft.ebookb4.models.enums;
+package kg.peaksoft.ebookb4.db.models.enums;
 
 public enum Language {
     ENGLISH, RUSSIAN, KYRGYZ;
