+++ src/main/java/kg/peaksoft/ebookb4/db/models/others/Favorites.java
@@ -1,6 +1,7 @@
-package kg.peaksoft.ebookb4.models.others;
+package kg.peaksoft.ebookb4.db.models.others;
 
-import kg.peaksoft.ebookb4.models.bookClasses.Book;
+import kg.peaksoft.ebookb4.db.models.bookClasses.Book;
+import kg.peaksoft.ebookb4.db.models.userClasses.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,18 +10,23 @@ import lombok.Setter;
 import javax.persistence.*;
 
 @Entity
-@Getter @Setter
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class FileSources {
+public class Favorites {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "generator")
     @SequenceGenerator(name = "generator", sequenceName = "generator", allocationSize = 1)
-    private Long id;
-
-    private String images;
+    @Column(name = "favorite_id")
+    private Long favoriteId;
 
     @ManyToOne
     @JoinColumn(name = "book_id")
     private Book book;
+
+    @OneToOne
+    @JoinColumn(name = "user_id")
+    private User User;
+
 }
