+++ src/main/java/kg/peaksoft/ebookb4/db/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package kg.peaksoft.ebookb4.repository;
+package kg.peaksoft.ebookb4.db.repository;
 
-import kg.peaksoft.ebookb4.models.userClasses.User;
+import kg.peaksoft.ebookb4.db.models.userClasses.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
