+++ src/main/java/kg/peaksoft/ebookb4/db/service/BookService.java
@@ -1,7 +1,7 @@
-package kg.peaksoft.ebookb4.service;
+package kg.peaksoft.ebookb4.db.service;
 
 import kg.peaksoft.ebookb4.dto.request.BookRequest;
-import kg.peaksoft.ebookb4.models.bookClasses.Book;
+import kg.peaksoft.ebookb4.db.models.bookClasses.Book;
 import org.springframework.http.ResponseEntity;
 
 import java.util.List;
