+++ src/main/java/kg/peaksoft/ebookb4/db/service/ClientService.java
@@ -1,4 +1,4 @@
-package kg.peaksoft.ebookb4.service;
+package kg.peaksoft.ebookb4.db.service;
 
 import kg.peaksoft.ebookb4.dto.request.SignupRequestClient;
 import org.springframework.http.ResponseEntity;
