+++ src/main/java/kg/peaksoft/ebookb4/db/service/impl/BookServiceImpl.java
@@ -1,19 +1,19 @@
-package kg.peaksoft.ebookb4.service.impl;
+package kg.peaksoft.ebookb4.db.service.impl;
 
+import kg.peaksoft.ebookb4.db.service.BookService;
 import kg.peaksoft.ebookb4.dto.mapper.BookMapper;
 import kg.peaksoft.ebookb4.dto.request.BookRequest;
 import kg.peaksoft.ebookb4.dto.response.MessageResponse;
 import kg.peaksoft.ebookb4.exceptions.NotFoundException;
-import kg.peaksoft.ebookb4.models.bookClasses.AudioBook;
-import kg.peaksoft.ebookb4.models.bookClasses.Book;
-import kg.peaksoft.ebookb4.models.bookClasses.ElectronicBook;
-import kg.peaksoft.ebookb4.models.bookClasses.PaperBook;
-import kg.peaksoft.ebookb4.models.enums.BookType;
-import kg.peaksoft.ebookb4.models.enums.Genre;
-import kg.peaksoft.ebookb4.models.enums.Language;
-import kg.peaksoft.ebookb4.repository.BookRepository;
-import kg.peaksoft.ebookb4.repository.UserRepository;
-import kg.peaksoft.ebookb4.service.BookService;
+import kg.peaksoft.ebookb4.db.models.bookClasses.AudioBook;
+import kg.peaksoft.ebookb4.db.models.bookClasses.Book;
+import kg.peaksoft.ebookb4.db.models.bookClasses.ElectronicBook;
+import kg.peaksoft.ebookb4.db.models.bookClasses.PaperBook;
+import kg.peaksoft.ebookb4.db.models.enums.BookType;
+import kg.peaksoft.ebookb4.db.models.enums.Genre;
+import kg.peaksoft.ebookb4.db.models.enums.Language;
+import kg.peaksoft.ebookb4.db.repository.BookRepository;
+import kg.peaksoft.ebookb4.db.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
