+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/ebook
 spring.datasource.username=postgres
-spring.datasource.password=54321
+spring.datasource.password=1992
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL10Dialect
 spring.jpa.show-sql=true
